requirements.txt
asgiref==3.3.1
Django==3.0.7 # d'origine 3.1.7
django-debug-toolbar==2.2
djangorestframework==3.11.0 # d'origine 3.12.2
mysql-connector==2.2.9
mysqlclient==1.4.6
nodeenv==1.5.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
